\name{aphid_cage_exp-package}
\alias{aphid_cage_exp-package}
\alias{aphid_cage_exp}
\docType{package}
\title{\packageTitle{aphid_cage_exp}}
\description{\packageDescription{aphid_cage_exp}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{aphid_cage_exp}
  \packageIndices{aphid_cage_exp}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{aphid_cage_exp}
  
  Maintainer: \packageMaintainer{aphid_cage_exp}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
