% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dev_times}
\alias{dev_times}
\title{Development times for aphids and wasps.}
\format{A list with two items:
\describe{
  \item{instar_days}{Number of days per instar, for low (\code{lowT}) and 
        high (\code{highT}) temperatures (20º and 27º C).}
  \item{mum_days}{1x2 matrix with the number of days per stage of an aphid 
                  being parasitized: living and dead ("mummy"), respectively.}
}}
\source{
\url{http://doi.wiley.com/10.1890/13-1933.1}
}
\usage{
dev_times
}
\description{
Development times for aphids and wasps.
}
\keyword{datasets}
