% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{environ}
\alias{environ}
\title{Parameters associated with environmental effects and stochasticity.}
\format{A list of length 9:
\describe{
  \item{harvest_surv}{
      Numeric vector of length 1 for aphid survival rate at harvesting.
  }
  \item{disp_aphid}{
      Numeric vector of length 1 for dispersal rates between fields for aphids, 
      adult wasps.
  }
  \item{disp_wasp}{
      Numeric vector of length 1 for dispersal rates between fields for aphids, 
      adult wasps.
  }
  \item{pred_rate}{
      Numeric vector of length 1 for predation rate for aphids and non-adult wasps.
  }
  \item{cycle_length}{
      Numeric vector of length 1 for time between harvests (typical for alfalfa).
  }
  \item{disp_start}{
    List of length 2, each item of which contains a 1-length numeric vector 
    indicating the day at which aphids begin dispersing for 20ºC  (\code{lowT}) 
    and 27ºC (\code{highT}).
    It's assumed that only adults are dispersing, but the day at which 
    this occurs depends on how quickly the aphids are developing.
  }
  \item{sigma_x}{
    Numeric vector of length 1, indicating environmental std dev for aphids.
  }
  \item{sigma_y}{
    Numeric vector of length 1, indicating environmental std dev for wasps.
  }
  \item{rho}{
    Numeric vector of length 1, indicating environmental correlation among instars.
  }
}}
\source{
\url{http://doi.wiley.com/10.1890/13-1933.1}
}
\usage{
environ
}
\description{
Parameters associated with environmental effects and stochasticity.
}
\keyword{datasets}
