% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{leslie_matrix}
\alias{leslie_matrix}
\title{Create Leslie matrix from aphid info}
\usage{
leslie_matrix(instar_days, surv_juv, surv_adult, repro)
}
\arguments{
\item{instar_days}{Integer vector of the number of stages (days) per aphid instar.}

\item{surv_juv}{Single numeric of daily juvenile survival.}

\item{surv_adult}{Numeric vector of aphid adult survivals by stage.}

\item{repro}{Numeric vector of aphid reproductive rates by stage.}
}
\description{
Create Leslie matrix from aphid info
}
