% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability.R
\name{make_sim_obj}
\alias{make_sim_obj}
\title{Create SimPatches object.}
\usage{
make_sim_obj(.n_pops, .n_patches, .harvest_periods = NULL,
  .harvest_offsets = 0, aphid_density_0 = NULL, aphid_surv_juv = NULL,
  aphid_surv_adult = NULL, aphid_repro = NULL, K = NULL,
  disp_aphid = NULL, attack_surv = NULL, no_error = TRUE, ...)
}
\arguments{
\item{.n_pops}{Number of aphid lines.}

\item{.n_patches}{Number of patches.}

\item{.harvest_periods}{Period of harvest cycle. Defaults to 
\code{sappp::environ$cycle_length}.}

\item{.harvest_offsets}{Offset for harvest cycle. Defaults to 0.}

\item{aphid_density_0}{Starting aphid densit(ies). Defaults to \code{NULL}.}

\item{aphid_surv_juv}{Aphid juvenile survival. Defaults to \code{NULL}.}

\item{aphid_surv_adult}{Vector of adult aphid survivals. Defaults to \code{NULL}.}

\item{aphid_repro}{Vector of aphid reproductive rates. Defaults to \code{NULL}.}

\item{K}{Aphid density dependency. Defaults to \code{NULL}.}

\item{disp_aphid}{Aphid dispersal rates. Defaults to \code{NULL}.}

\item{attack_surv}{Aphid wasp-attack survival rates, for singly and multiply 
infected aphids. Defaults to \code{NULL}.}

\item{no_error}{Boolean for whether to include no process error. Defaults to 
\code{TRUE}.}

\item{...}{Other parameters associated with aphid lines and their infecting wasps.
See \code{\link{all_pop_lists}} for info on how to provide these if you wish
to vary by line or patch, and 
see \code{vignette(parameters)} for the parameters available.}
}
\value{
An SimPatches object.
}
\description{
For parameters \code{aphid_density_0}, \code{aphid_surv_juv}, \code{aphid_surv_adult},
\code{aphid_repro}, \code{K}, \code{disp_aphid}, and \code{attack_surv}, 
you should provide a list containing, at minimum, the standard deviation of the 
variability that you desire.
You can also provide a mean value from which deviates will be created.
}
\details{
For \code{.harvest_periods} and {.harvest_offsets}, provide either a single integer
or a vector of a length that is a factor of \code{.n_patches}.
}
\examples{

so <- make_sim_obj(
    .n_pops = 10, .n_patches = 8,
    .harvest_periods = 7,
    .harvest_offsets = c(0, 3),
    aphid_density_0 = NULL,
    aphid_surv_juv = list(.sd_pops = 2),
    aphid_surv_adult = list(.sd_pops = 1),
    aphid_repro = list(.sd_pops = 2),
    K = list(.sd_pops = 1),
    disp_aphid = list(.sd_pops = 1),
    attack_surv = NULL,
    no_error = TRUE)
sim <- so$simulate(630)
sim


}
