% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{populations}
\alias{populations}
\title{Population rates and starting values for aphids and wasps.}
\format{A list with ten items:
\describe{
  \item{surv_juv} {
      List of length two, each item a length-of-one numeric vector containing
      juvenile daily survival for aphid lines with low and high survivals
      (\code{low} and \code{high}).
  }
  \item{surv_adult} {
      List of length two, each item a 1x200 matrix containing
      adults daily survival for aphid lines with low and high survivals
      (\code{low} and \code{high}).
      Although each matrix has 200 columns, most of them are filled with zeros.
  }
  \item{repro} {
      List of length two, each item a 1x200 matrix containing
      daily reproduction for aphid lines with low and high reproduction
      (\code{low} and \code{high}).
      Although each matrix has 200 columns, most of them are filled with zeros.
  }
  \item{K}{
    A single number representing aphid density dependence.
  }
  \item{K_y}{
    A single number representing parasitized aphid density dependence.
  }
  \item{s_y}{
    A single number representing parasitoid adult daily survival.
  }
  \item{sex_ratio}{
    A single number representing proportion of female wasps.
  }
  \item{aphids_0}{
    A single number representing initial density of aphids.
  }
  \item{wasps_0}{
    A single number representing initial densities of wasps.
  }
  \item{prop_resist}{
    A single number representing proportion of resistant clones.
  }
}}
\source{
\url{http://doi.wiley.com/10.1890/13-1933.1}
}
\usage{
populations
}
\description{
Population rates and starting values for aphids and wasps.
}
\keyword{datasets}
