% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sappp.R
\docType{package}
\name{sappp}
\alias{sappp}
\alias{sappp-package}
\title{sappp: Simulating Aphid and Parasitoid Populations... with Predation.}
\description{
sappp: Simulating Aphid and Parasitoid Populations... with Predation.
}
\section{safepug functions}{

}

